♻️ [refactor] : 시작 파일 통합 및 유틸 모듈화
