+++ chat_server/package.json
@@ -1,7 +1,7 @@
 {
 	"scripts": {
-		"start": "node dist/index.js",
-		"dev": "nodemon --watch \"*.ts\" --exec \"ts-node\" --files ./index.ts",
+		"start": "node dist/app.js",
+		"dev": "nodemon --watch \"*.ts\" --exec \"ts-node\" --files ./app.ts",
 		"build": "tsc -p .",
 		"format:check": "npm run format:check:partial -w=community-board --include-workspace-root -- chat_server shared",
 		"format:write": "npm run format:write:partial -w=community-board --include-workspace-root -- chat_server shared"
@@ -16,7 +16,6 @@
 		"socket.io": "^4.7.5"
 	},
 	"devDependencies": {
-		"@socket.io/admin-ui": "^0.5.1",
 		"@types/express": "^4.17.21",
 		"@types/node": "^22.2.0",
 		"@types/jsonwebtoken": "^9.0.6",
